+++ gframe/sound_manager.cpp
@@ -84,7 +84,7 @@ void SoundManager::RefreshChantsList() {
 		for (auto& file : Utils::FindFiles(searchPath, { EPRO_TEXT("mp3"), EPRO_TEXT("ogg"), EPRO_TEXT("wav"), EPRO_TEXT("flac") })) {
 			auto scode = Utils::GetFileName(searchPath + EPRO_TEXT("/") + file);
 			try {
-				unsigned int code = std::stoi(scode);
+				uint32_t code = static_cast<uint32_t>(std::stoul(scode));
 				auto key = std::make_pair(chantType.first, code);
 				if (code && !ChantsList.count(key))
 					ChantsList[key] = working_dir + "/" + Utils::ToUTF8IfNeeded(searchPath + EPRO_TEXT("/") + file);
@@ -140,7 +140,7 @@ void SoundManager::PlayBGM(BGM scene, bool loop) {
 	}
 #endif
 }
-bool SoundManager::PlayChant(CHANT chant, unsigned int code) {
+bool SoundManager::PlayChant(CHANT chant, uint32_t code) {
 #ifdef BACKEND
 	if(!soundsEnabled) return false;
 	auto key = std::make_pair(chant, code);
